与前端初步对接的版本

+++ src/main/java/com/example/pojo/PageBean.java
@@ -13,7 +13,7 @@ import java.util.List;
 @NoArgsConstructor
 @AllArgsConstructor
 public class PageBean {
-    private Long total; //总记录数
+    private Integer total; //总记录数
     private List rows; //当前页数据列表
-    private Long pageCount; //总页数
+    private Integer pageCount; //总页数
 }
